requirements.txt
streamlit
pandas
plotly
sqlalchemy
openpyxl
lxml
google-auth-oauthlib
msal
streamlit-lottie
pypdf
python-pptx
bcrypt
python-docx
